+++ Glados/src/Platform/OpenGL/OpenGLBuffer.h
@@ -25,14 +25,14 @@ namespace Glados {
 		BufferLayout m_Layout;
 		Ref<IndexBuffer> m_IndexBuffer;
 	public:
-		OpenGLVertexBuffer(uint32_t size);
-		OpenGLVertexBuffer(float* vertices, uint32_t size);
+		OpenGLVertexBuffer(size_t size);
+		OpenGLVertexBuffer(float* vertices, size_t size);
 
 		virtual void Bind() const override;
 		virtual void Unbind() const override;
 
-		virtual void SetData(const void* data, uint32_t offset, uint32_t size) override;
-		virtual float GetData(uint32_t offset, uint32_t size) const override;
+		virtual void SetData(const void* data, uint32_t offset, size_t size) override;
+		virtual float GetData(uint32_t offset, size_t size) const override;
 
 		virtual void SetLayout(const BufferLayout& layout) override;
 		virtual BufferLayout GetLayout() const override;
@@ -45,14 +45,14 @@ namespace Glados {
 		uint32_t m_RendererID;
 		BufferLayout m_Layout;
 	public:
-		OpenGLUniformBuffer(uint32_t size);
-		OpenGLUniformBuffer(float* vertices, uint32_t size);
+		OpenGLUniformBuffer(size_t size);
+		OpenGLUniformBuffer(float* vertices, size_t size);
 
 		virtual void Bind() const override;
 		virtual void Unbind() const override;
 
-		virtual void SetData(const void* data, uint32_t offset, unsigned int size) override;
-		virtual float GetData(uint32_t offset, uint32_t size) const override;
+		virtual void SetData(const void* data, uint32_t offset, size_t size) override;
+		virtual float GetData(uint32_t offset, size_t size) const override;
 
 		virtual void SetLayout(const BufferLayout& layout) override;
 		virtual BufferLayout GetLayout() const override;
